Use the correct PSI element type in completion LookupItems to properly display a summary of bibtex entries in the documentation popup
