+++ src/nl/hannahsten/texifyidea/completion/LatexBibliographyReferenceProvider.kt
@@ -23,7 +23,7 @@ object LatexBibliographyReferenceProvider : CompletionProvider<CompletionParamet
                         val lookupStrings = LinkedList(bibtexEntry.authors)
                         lookupStrings.add(bibtexEntry.title)
                         LookupElementBuilder.create(bibtexEntry.identifier)
-                            .withPsiElement(bibtexEntry)
+                            .withPsiElement(bibtexEntry.id)
                             .withPresentableText(bibtexEntry.title)
                             .bold()
                             .withInsertHandler(LatexReferenceInsertHandler())
