+++ test/nl/hannahsten/texifyidea/reference/LatexLabelCompletionTest.kt
@@ -6,66 +6,6 @@ import nl.hannahsten.texifyidea.file.LatexFileType
 import org.junit.Test
 
 class LatexLabelCompletionTest : BasePlatformTestCase() {
-    override fun getTestDataPath(): String {
-        return "test/resources/completion/cite"
-    }
-
-    @Throws(Exception::class)
-    override fun setUp() {
-        super.setUp()
-    }
-
-    @Test
-    fun testCompleteLatexReferences() {
-        // when
-        runCompletion()
-        val result = myFixture.lookupElements!!
-
-        // then
-        assertEquals(3, result.size)
-        val entry1 = result.first { l -> l!!.lookupString == "Evans2015" }!!
-        assertTrue(entry1.allLookupStrings.contains("Evans, Isaac"))
-        assertTrue(entry1.allLookupStrings.contains("Evans2015"))
-        assertTrue(entry1.allLookupStrings.contains("Missing the Point(er): On the Effectiveness of Code Pointer Integrity"))
-    }
-
-    @Test
-    fun testCompletionResultsLowerCase() {
-        // when
-        runCompletion()
-        val result = myFixture.lookupElementStrings
-
-        // then
-        assertEquals(1, result?.size)
-        assertTrue(result?.contains("Muchnick1997") == true)
-    }
-
-    @Test
-    fun testCompletionResultsSecondEntry() {
-        // when
-        runCompletion()
-        val result = myFixture.lookupElementStrings
-
-        // then
-        assertEquals(3, result?.size)
-        assertTrue(result?.contains("Muchnick1997") == true)
-        assertTrue(result?.contains("Evans2015") == true)
-        assertTrue(result?.contains("Burow2016") == true)
-    }
-
-    private fun runCompletion() {
-        myFixture.configureByFiles("${getTestName(false)}.tex", "bibtex.bib")
-
-        // when
-        myFixture.complete(CompletionType.BASIC)
-    }
-
-    @Test
-    fun testCompleteBibtexWithCorrectCase() {
-        // Using the following failed sometimes
-        val testName = getTestName(false)
-        myFixture.testCompletion("${testName}_before.tex", "${testName}_after.tex", "$testName.bib")
-    }
 
     @Test
     fun testLabelReferenceCompletion() {
