+++ test/resources/completion/cite/BibtexEntryDocumentation.tex
@@ -1,0 +1,13 @@
+
+% Preamble
+\documentclass[11pt]{article}
+\usepackage{natbib}
+
+% Document
+\begin{document}
+
+    \cite{isaac<caret>}
+    \bibliography{bibtex}
+    \bibliographystyle{plain}
+
+\end{document}
